// src/integrations/supabase/types.ts
export type Database = {
  public: {
// ...
    Enums: {
      app_language: "en" | "es" | "fr" | "de";
      user_role: "user" | "admin";
      app_theme: "light" | "dark" | "system";
      sop_category:
        | "Marketing"
        | "HR"
        | "Operations"
        | "Finance"
        | "Customer Service"
        | "IT"
        | "Sales"
        | "Quality Assurance";
      subscription_plan: "Free" | "Pro" | "Team";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DefaultSchema = Database[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;